+++ src/Services/.BaseService/AppyNox.Services.Base.API/Filters/JwtTokenValidateAttribute.cs
@@ -7,14 +7,14 @@ namespace AppyNox.Services.Base.API.Filters
 {
     [AttributeUsage(AttributeTargets.Class | AttributeTargets.Method, Inherited = false, AllowMultiple = false)]
     [Obsolete("JwtTokenValidateAttribute is deprecated. Use NoxJwtAuthenticationHandler instead. This class will be removed in v1.0.5")]
-    public class JwtTokenValidateAttribute : Attribute, IAuthorizationFilter
+    public class JwtTokenValidateAttribute : Attribute, IAsyncAuthorizationFilter
     {
         #region [ Public Methods ]
 
-        public virtual void OnAuthorization(AuthorizationFilterContext context)
+        async Task IAsyncAuthorizationFilter.OnAuthorizationAsync(AuthorizationFilterContext context)
         {
             string token = context.HttpContext.Request.Headers[HeaderNames.Authorization].ToString().Replace("Bearer ", "");
-            if (string.IsNullOrWhiteSpace(token) || (context.HttpContext.RequestServices.GetService(typeof(INoxTokenManager)) as INoxTokenManager)?.VerifyToken(token) == false)
+            if (string.IsNullOrWhiteSpace(token) || await (context.HttpContext.RequestServices.GetService(typeof(INoxTokenManager)) as INoxTokenManager)?.VerifyToken(token)! == false)
             {
                 context.Result = new UnauthorizedResult();
             }