+++ src/Services/LicenseService/AppyNox.Services.License.Application/Validators/Product/Update/ProductUpdateCreateValidator.cs
@@ -1,8 +1,9 @@
 ﻿using AppyNox.Services.Base.Application.Validators;
 using AppyNox.Services.License.Application.Dtos.ProductDtos.Models.Base;
+using AppyNox.Services.License.Application.Validators.Product.Create;
 using FluentValidation;
 
-namespace AppyNox.Services.License.Application.Validators.Product.Create
+namespace AppyNox.Services.License.Application.Validators.Product.Update
 {
     public class ProductUpdateCreateValidator : DtoValidatorBase<ProductSimpleUpdateDto>
     {