+++ src/Services/LicenseService/AppyNox.Services.License.WebAPI/Controllers/v1/LicensesController.cs
@@ -1,4 +1,5 @@
-﻿using AppyNox.Services.Base.API.Helpers;
+﻿using AppyNox.Services.Base.API.Constants;
+using AppyNox.Services.Base.API.Helpers;
 using AppyNox.Services.Base.API.ViewModels;
 using AppyNox.Services.Base.Application.MediatR.Commands;
 using AppyNox.Services.Base.Infrastructure.Repositories.Common;
@@ -6,7 +7,6 @@ using AppyNox.Services.License.Application.MediatR.Commands;
 using AppyNox.Services.License.Domain.Entities;
 using AppyNox.Services.License.WebAPI.Permission;
 using Asp.Versioning;
-using AutoWrapper.Wrappers;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -14,7 +14,7 @@ using Microsoft.AspNetCore.Mvc;
 namespace AppyNox.Services.License.WebAPI.Controllers.v1
 {
     [ApiController]
-    [ApiVersion("1.0")]
+    [ApiVersion(NoxVersions.v1_0)]
     [Route("api/[controller]")]
     public class LicensesController(IMediator mediator) : Controller
     {
@@ -28,16 +28,16 @@ namespace AppyNox.Services.License.WebAPI.Controllers.v1
 
         [HttpGet]
         [Authorize(Permissions.Licenses.View)]
-        public async Task<ApiResponse> GetAll([FromQuery] QueryParametersViewModel queryParameters)
+        public async Task<IActionResult> GetAll([FromQuery] QueryParametersViewModel queryParameters)
         {
-            return new ApiResponse(await _mediator.Send(new GetAllEntitiesQuery<LicenseEntity>(queryParameters)));
+            return Ok(await _mediator.Send(new GetAllEntitiesQuery<LicenseEntity>(queryParameters)));
         }
 
         [HttpGet("{id}")]
         [Authorize(Permissions.Licenses.View)]
-        public async Task<ApiResponse> GetById(Guid id, [FromQuery] QueryParametersViewModel queryParameters)
+        public async Task<IActionResult> GetById(Guid id, [FromQuery] QueryParametersViewModel queryParameters)
         {
-            return new ApiResponse(await _mediator.Send(new GetEntityByIdQuery<LicenseEntity>(id, queryParameters)));
+            return Ok(await _mediator.Send(new GetEntityByIdQuery<LicenseEntity>(id, queryParameters)));
         }
 
         [HttpPost]
@@ -83,7 +83,7 @@ namespace AppyNox.Services.License.WebAPI.Controllers.v1
         #region [ Custom Endpoints ]
 
         [HttpGet("validate/{licenseKey}")]
-        public async Task<ActionResult> ValidateLicenseKey(string licenseKey)
+        public async Task<IActionResult> ValidateLicenseKey(string licenseKey)
         {
             var response = await _mediator.Send(new ValidateLicenseKeyCommand(licenseKey));