+++ src/Services/LicenseService/AppyNox.Services.License.WebAPI/Controllers/v1/ProductsController.cs
@@ -1,17 +1,17 @@
-﻿using AppyNox.Services.Base.API.Helpers;
+﻿using AppyNox.Services.Base.API.Constants;
+using AppyNox.Services.Base.API.Helpers;
 using AppyNox.Services.Base.API.ViewModels;
 using AppyNox.Services.Base.Application.MediatR.Commands;
 using AppyNox.Services.Base.Infrastructure.Repositories.Common;
 using AppyNox.Services.License.Domain.Entities;
 using Asp.Versioning;
-using AutoWrapper.Wrappers;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
 
 namespace AppyNox.Services.License.WebAPI.Controllers.v1
 {
     [ApiController]
-    [ApiVersion("1.0")]
+    [ApiVersion(NoxVersions.v1_0)]
     [Route("api/[controller]")]
     public class ProductsController(IMediator mediator) : Controller
     {
@@ -24,15 +24,15 @@ namespace AppyNox.Services.License.WebAPI.Controllers.v1
         #region [ CRUD Endpoints ]
 
         [HttpGet]
-        public async Task<ApiResponse> GetAll([FromQuery] QueryParametersViewModel queryParameters)
+        public async Task<IActionResult> GetAll([FromQuery] QueryParametersViewModel queryParameters)
         {
-            return new ApiResponse(await _mediator.Send(new GetAllEntitiesQuery<ProductEntity>(queryParameters)));
+            return Ok(await _mediator.Send(new GetAllEntitiesQuery<ProductEntity>(queryParameters)));
         }
 
         [HttpGet("{id}")]
-        public async Task<ApiResponse> GetById(Guid id, [FromQuery] QueryParametersViewModel queryParameters)
+        public async Task<IActionResult> GetById(Guid id, [FromQuery] QueryParametersViewModel queryParameters)
         {
-            return new ApiResponse(await _mediator.Send(new GetEntityByIdQuery<ProductEntity>(id, queryParameters)));
+            return Ok(await _mediator.Send(new GetEntityByIdQuery<ProductEntity>(id, queryParameters)));
         }
 
         [HttpPost]